moved all firebase functions to a single file
